yank.sh: Remove tmux passthrough
Tmux passthrough was previously used for bypassing tmux handling of
escape codes as it didn't know about OSC52. This is counter-productive
now as bypassing tmux means it doesn't get the updated contents of the
clipboard. Disable this passthrough and add DEBUG_YANK variable to print
out information when debugging the failure of yank.sh.